feat: Add dynamic choices to Choice Prompt (#1777)
* Update validation and choice array ui

* Updated string array

* Updated regex pattern

* Updated array styles

* Style cleanup

* Fixed string array test

* Fixed ObjectArray tests

* Updated request header object

* Small change to custom object field

* Fixed styling

* Updated dialogs

* Fix margin issue

* Updated callback dependencies

* Added error message

* Updated callback dependencies

* Fixed margin issues

* Fixed styles

* Fixed transparent border issue

* Created Dyncamic and Static Choices components

* Added Static Choices

* Added ExpressionWidget to Dynamic Choice and updated IChoice type
