+++ Composer/packages/extensions/obiformeditor/src/Form/fields/PromptField/ChoiceInput/index.tsx
@@ -11,11 +11,13 @@ import { ChoiceInput, IChoiceOption } from '@bfc/shared';
 import { PromptFieldChangeHandler, GetSchema } from '../types';
 import { CheckboxWidget } from '../../../widgets';
 import { field } from '../styles';
+import { FormContext } from '../../../types';
 
 import { Choices } from './Choices';
 import { ChoiceOptions } from './ChoiceOptions';
 
 interface ChoiceInputSettingsProps extends FieldProps<ChoiceInput> {
+  formContext: FormContext;
   onChange: PromptFieldChangeHandler;
   getSchema: GetSchema;
 }
@@ -31,6 +33,7 @@ export const ChoiceInputSettings: React.FC<ChoiceInputSettingsProps> = props => 
   return (
     <Fragment>
       <Choices
+        formContext={formContext}
         formData={formData.choices}
         schema={getSchema('choices')}
         onChange={onChange('choices')}
